+++ nanojit/Assembler.h
@@ -202,7 +202,7 @@ namespace nanojit
 			void		setError(AssmError e) { _err = e; }
 			void		setCallTable(const CallInfo *functions);
 			void		pageReset();
-			int32_t		codeBytes();
+			size_t		codeBytes();
 			Page*		handoverPages(bool exitPages=false);
 
 			debug_only ( void		pageValidate(); )
